[refactor/#7] Single Activity Architecture 구조로 수정
