+++ app/src/main/AndroidManifest.xml
@@ -17,15 +17,6 @@
         tools:targetApi="31">
         <activity
             android:name=".ui.MainActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.mypage.screen.MyActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.onboarding.screen.SignUpActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.onboarding.screen.SignInActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ATSOPTANDROID">
