+++ app/src/main/java/org/sopt/at/ui/MainActivity.kt
@@ -7,6 +7,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -17,6 +18,9 @@ import org.sopt.at.ui.history.screen.HistoryScreen
 import org.sopt.at.ui.home.screen.HomeScreen
 import org.sopt.at.ui.home.viewmodel.ContentViewModel
 import org.sopt.at.ui.live.screen.LiveScreen
+import org.sopt.at.ui.mypage.screen.MyScreen
+import org.sopt.at.ui.onboarding.screen.SignInScreen
+import org.sopt.at.ui.onboarding.screen.SignUpScreen
 import org.sopt.at.ui.search.screen.SearchScreen
 import org.sopt.at.ui.shorts.screen.ShortsScreen
 import org.sopt.at.ui.theme.TvingTheme
@@ -30,16 +34,38 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             val navController = rememberNavController()
+            val currentBackStackEntry =
+                navController.currentBackStackEntryFlow.collectAsState(initial = navController.currentBackStackEntry)
+            val currentRoute = currentBackStackEntry.value?.destination?.route  // 현재 화면
+            val showBottomNavBar = when (currentRoute) {
+                NavRoutes.SignIn.route, NavRoutes.SignUp.route -> false
+                else -> true
+            }
 
             TvingTheme {
                 Scaffold(
-                    bottomBar = { BottomNavBar(navController) }
+                    bottomBar = {
+                        if (showBottomNavBar) {
+                            BottomNavBar(navController)
+                        }
+                    }
                 ) { innerPadding ->
                     NavHost(
                         navController = navController,
-                        startDestination = NavRoutes.Home.route,
+                        startDestination = NavRoutes.SignIn.route,
                         modifier = Modifier.padding(innerPadding)
                     ) {
+                        composable(NavRoutes.SignIn.route) {
+                            SignInScreen(navController = navController)
+                        }
+                        composable(NavRoutes.SignUp.route) {
+                            SignUpScreen(navController = navController)
+                        }
+
+                        composable(NavRoutes.Mypage.route) {
+                            MyScreen(navController = navController)
+                        }
+
                         composable(NavRoutes.Home.route) {
                             HomeScreen(contentViewModel)
                         }
