+++ app/src/main/java/org/sopt/at/ui/mypage/screen/MyScreen.kt
@@ -1,10 +1,5 @@
 package org.sopt.at.ui.mypage.screen
 
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -17,54 +12,28 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
 import org.sopt.at.R
 import org.sopt.at.ui.common.component.ButtonComponent
 import org.sopt.at.ui.common.component.TopAppBarComponent
-import org.sopt.at.ui.onboarding.screen.SignInActivity
-import org.sopt.at.ui.theme.TvingTheme
 import org.sopt.at.ui.theme.TvingTheme.colors
 
-class MyActivity : ComponentActivity() {
-    private var loggedInId: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // 로그인에서 넘어온 정보 받기
-        loggedInId = intent.getStringExtra("id")
-
-        enableEdgeToEdge()
-        setContent {
-            TvingTheme {
-                loggedInId?.let {
-                    MyScreen(
-                        profileId = it
-                    )
-                }
-            }
-        }
-    }
-}
-
 @Composable
-fun MyScreen(
-    profileId: String
-) {
-    val context = LocalContext.current
-    val intent = Intent(context, SignInActivity::class.java).apply {
-        flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
-    }
+fun MyScreen(navController: NavController) {
+    var profileId by remember { mutableStateOf("") }
 
     Column(
         modifier = Modifier
@@ -110,16 +79,8 @@ fun MyScreen(
             containerColor = colors.BasicBlack,
             contentColor = colors.Gray3,
             text = stringResource(R.string.logout),
-            onClick = { context.startActivity(intent) },
+            onClick = { /* TODO: SignInScreen으로 이동 */ },
             strokeColor = colors.Gray3
         )
     }
-}
-
-@Preview
-@Composable
-fun MyScreenPreview() {
-    MyScreen(
-        profileId = "홍길동"
-    )
 }
